requirements.txt
streamlit
pdf2image
pillow
imagehash
openpyxl
pandas
pdfplumber
pytesseract
